:recycle: move oauth_url and Undefined class to public.py; update imports
